requirements.txt
matplotlib==3.5.1
numpy==1.21.6
pillow==9.5.0
python==3.7.16
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111